refactor(crafter): refactored sentenzier crafter
