+++ crafters/numeric/ArrayBytesReader/README.md
@@ -1,0 +1,3 @@
+# ArrayBytesReader
+
+converts a byte stream into a numpy array and save to the Document.
