+++ crafters/numeric/ArrayBytesReader/config.yml
@@ -1,0 +1,3 @@
+!ArrayBytesReader
+with:
+  {}
