+++ crafters/numeric/ArrayBytesReader/manifest.yml
@@ -1,0 +1,11 @@
+manifest_version: 1
+name: ArrayBytesReader
+type: crafter
+description: converts a byte stream into a numpy array and save to the Document.
+author: Jina AI Dev-Team (dev-team@jina.ai)
+url: https://jina.ai
+vendor: Jina AI Limited
+documentation: https://github.com/jina-ai/jina-hub
+version: 0.0.1
+license: apache-2.0
+keywords: [Some keywords to describe the executor, separated by commas]
