+++ crafters/numeric/ArrayStringReader/Dockerfile
@@ -1,0 +1,13 @@
+FROM jinaai/jina:devel
+
+# setup the workspace
+COPY . /workspace
+WORKDIR /workspace
+
+# install the third-party requirements
+RUN pip install -r requirements.txt
+
+# for testing the image
+RUN pip install pytest && pytest
+
+ENTRYPOINT ["jina", "pod", "--uses", "config.yml"]
