+++ crafters/numeric/ArrayStringReader/README.md
@@ -1,0 +1,3 @@
+# ArrayStringReader
+
+convertsthe string of numbers into a numpy array and save to the Document.
