+++ crafters/numeric/ArrayStringReader/config.yml
@@ -1,0 +1,3 @@
+!ArrayStringReader
+with:
+  {}
